+++ src/SeerVariableManagerWidget.cpp
@@ -3,6 +3,7 @@
 #include "QHContainerWidget.h"
 #include <QtWidgets/QToolButton>
 #include <QtGui/QIcon>
+#include <QtCore/QSettings>
 #include <QtCore/QDebug>
 
 SeerVariableManagerWidget::SeerVariableManagerWidget (QWidget* parent) : QWidget(parent) {
@@ -39,9 +40,14 @@ SeerVariableManagerWidget::SeerVariableManagerWidget (QWidget* parent) : QWidget
 
     tabWidget->setCornerWidget(hcontainer, Qt::TopRightCorner);
 
+    // Restore tab ordering.
+    readSettings();
+
     // Connect things.
-    QObject::connect(refreshToolButton, &QToolButton::clicked,     this,  &SeerVariableManagerWidget::handleRefreshToolButtonClicked);
-    QObject::connect(helpToolButton,    &QToolButton::clicked,     this,  &SeerVariableManagerWidget::handleHelpToolButtonClicked);
+    QObject::connect(refreshToolButton,     &QToolButton::clicked,     this,  &SeerVariableManagerWidget::handleRefreshToolButtonClicked);
+    QObject::connect(helpToolButton,        &QToolButton::clicked,     this,  &SeerVariableManagerWidget::handleHelpToolButtonClicked);
+    QObject::connect(tabWidget->tabBar(),   &QTabBar::tabMoved,        this,  &SeerVariableManagerWidget::handleTabMoved);
+    QObject::connect(tabWidget->tabBar(),   &QTabBar::currentChanged,  this,  &SeerVariableManagerWidget::handleTabChanged);
 }
 
 SeerVariableManagerWidget::~SeerVariableManagerWidget () {
@@ -71,5 +77,93 @@ void SeerVariableManagerWidget::handleHelpToolButtonClicked () {
     help->loadFile(":/seer/resources/help/VariableRegisterInfoBrowser.md");
     help->show();
     help->raise();
+}
+
+void SeerVariableManagerWidget::handleTabMoved (int from, int to) {
+
+    Q_UNUSED(from);
+    Q_UNUSED(to);
+
+    writeSettings();
+}
+
+void SeerVariableManagerWidget::handleTabChanged (int index) {
+
+    Q_UNUSED(index);
+
+    writeSettings();
+}
+
+void SeerVariableManagerWidget::writeSettings () {
+
+    // Write tab order to settings.
+    QStringList tabs;
+
+    for (int i=0; i<tabWidget->tabBar()->count(); i++) {
+        tabs.append(tabWidget->tabBar()->tabText(i));
+    }
+
+    QString current = tabWidget->tabBar()->tabText(tabWidget->tabBar()->currentIndex());
+
+    //qDebug() << "Tabs" << tabs;
+    //qDebug() << "Current" << current;
+
+    QSettings settings;
+
+    settings.beginGroup("variablemanagerwindow"); {
+        settings.setValue("taborder",   tabs.join(','));
+        settings.setValue("tabcurrent", current);
+    } settings.endGroup();
+}
+
+void SeerVariableManagerWidget::readSettings () {
+
+    // Can't move things?
+    if (tabWidget->tabBar()->isMovable() == false) {
+        return;
+    }
+
+    // Read tab order from settings.
+    QSettings   settings;
+    QStringList tabs;
+    QString     current;
+
+    settings.beginGroup("variablemanagerwindow"); {
+        tabs    = settings.value("taborder").toString().split(',');
+        current = settings.value("tabcurrent").toString();
+    } settings.endGroup();
+
+    //qDebug() << "Tabs"    << tabs;
+    //qDebug() << "Current" << current;
+
+    // Move tabs to the requested order.
+    for (int i=0; i<tabs.count(); i++) {
+
+        QString tab = tabs[i];
+        int     tb  = -1;
+
+        for (int j=0; j<tabWidget->tabBar()->count(); j++) {
+            if (tabWidget->tabBar()->tabText(j) == tab) {
+                tb = j;
+                break;
+            }
+        }
+
+        if (tb != -1) {
+            tabWidget->tabBar()->moveTab(tb, i);
+        }
+    }
+
+    // Make a tab current.
+    if (current != "") {
+        for (int i=0; i<tabWidget->tabBar()->count(); i++) {
+            if (tabWidget->tabBar()->tabText(i) == current) {
+                tabWidget->setCurrentIndex(i);
+                break;
+            }
+        }
+    }else{
+        tabWidget->setCurrentIndex(0);
+    }
 }